Added support for writing
Fixed connected check not working for discoverServices.
Added NetworkError catch.
Added support to convert the data in a format ready for the api call.
Fixed request options to not include the not set (null) attributes.
Converted the uuids in request options to lower case since this is what the api expects.