Created new builders for manufacturer identifiers and default service and characteristic uuids.
These builders read in the previous version of the enums to mark the old names, and/ or removed entries as deprecated instead of just removing them.

Also get the latest information from the official Bluetooth bitbucket repository to make it a lot less labor-intensive to update the enum.

Added the official Bluetooth SIG id to the default service services and characteristics.

Used the new builder to generate `bluetooth_default_manufacturer_identifiers.dart` and `bluetooth_default_uuids.dart`.

Used the new `BluetoothDefaultManufacturerIdentifiers.manufacturerIdentifiers`, `BluetoothDefaultServiceUUIDS.services`, and `BluetoothDefaultCharacteristicUUIDS.characteristics` lists instead of the `.values` in the example. These lists don't contain any of the deprecated entries.